fix: correct DSP formulas across spectral, voice, pitch, temporal, cepstral
Accuracy corrections verified against librosa and closed-form references:
- spectral_rolloff cumulates magnitude (librosa parity, exact); flatness uses
  power spectrum; crest = max/mean; slope per-Hz with sample_rate
- YIN gains parabolic interpolation + true difference function (<0.3% error)
- voice_onset_time autocorrelation fixed (was irfft(rfft(x)) identity)
- CPP power-cepstrum 10*log10 with regression-subtracted peak picking
- vocal_tract_length quarter-wave c/(4*F1); harmonic_diff correct bin width
- LSP extraction rewritten (valid interleaved pairs); LPCC sign + index fix
- alpha_ratio/hammarberg/nasality frame+average (no longer truncate input)
- jitter_rap, shimmer_apq5/apq11, harmonic_to_noise_ratio_acf added
- temporal_centroid is MPEG-7 seconds; framewise variant kept under new name
- silence/length-1 input guards across ~15 functions (finite, no NaN)
- segmentation: seeded k-means, off-diagonal thumbnailing, scale-free silence
- explicit __all__ on feature modules; subpackage exports completed

+++ audiofeat/pitch/strength.py
@@ -2,7 +2,7 @@
 import torch
 from ..temporal.rms import frame_signal, hann_window
 
-def pitch_strength(x: torch.Tensor, fs: int, frame_length: int, hop_length: int, fmin: int = 50, fmax: int = 600):
+def pitch_strength(x: torch.Tensor, fs: int, frame_length: int, hop_length: int, fmin: float = 50.0, fmax: float = 600.0):
     """
     Computes the pitch strength using the autocorrelation method.
